Restructuring API controllers into their respective regions

+++ PowerLifting.Data/Entities/ProgramLogs/ProgramLogRepScheme.cs
@@ -7,7 +7,7 @@
         public int SetNo { get; set; }
         public int NoOfReps { get; set; }
         public decimal WeightLifted { get; set; }
-        public double? Percentage { get; set; }
+        public decimal? Percentage { get; set; }
         public string Comment { get; set; }
         public bool AMRAP { get; set; } //As many reps as possible for this set
         public bool Completed { get; set; }
